requirements.txt
streamlit
langchain
faiss-cpu
python-dotenv
openai>=1.0.0,<2.0.0
